+++ code/utils/table/TableSessionContext.tsx
@@ -6,10 +6,12 @@ import { LifecycleStage } from '@/types/form';
 import { TableRowHandle } from '@/ui/graphic/table/row/table-row';
 import HistoryModal from '@/ui/interaction/modal/history-modal';
 import { TableScrollDescriptor } from '@/hooks/table/useTableScroll';
+import { RegistryExportSettings } from '@/types/settings';
 
 export interface TableSessionState {
     activeRowId: string;
     recordType: string;
+    exports: RegistryExportSettings[];
     lifecycleStage: LifecycleStage;
     tableDescriptor: TableDescriptor;
     tableScrollDescriptor: TableScrollDescriptor
@@ -25,6 +27,7 @@ export const TableSessionContext = createContext<TableSessionState>(null);
 
 export const TableSessionContextProvider = ({
     recordType,
+    exports,
     lifecycleStage,
     tableDescriptor,
     tableScrollDescriptor,
@@ -34,6 +37,7 @@ export const TableSessionContextProvider = ({
     children,
 }: {
     recordType: string;
+    exports: RegistryExportSettings[];
     lifecycleStage: LifecycleStage;
     tableDescriptor: TableDescriptor;
     tableScrollDescriptor: TableScrollDescriptor
@@ -47,7 +51,7 @@ export const TableSessionContextProvider = ({
     const [activeRowId, setActiveRowId] = useState<string>("");
 
     return (
-        <TableSessionContext.Provider value={{ activeRowId, recordType, lifecycleStage, tableDescriptor, tableScrollDescriptor, rowRefs, addEntity, pricingType, setActiveRowId, setHistoryId, setIsOpenHistoryModal }}>
+        <TableSessionContext.Provider value={{ activeRowId, recordType, exports, lifecycleStage, tableDescriptor, tableScrollDescriptor, rowRefs, addEntity, pricingType, setActiveRowId, setHistoryId, setIsOpenHistoryModal }}>
             {children}
             {isOpenHistoryModal && historyId != "" &&
                 <HistoryModal